Fix bug with imageURL genertion

+++ build/deploy-functions.sh
@@ -3,6 +3,11 @@ cd dist
 
 for file in *.js; do
   filename="${file%%.*}"
+  if [[ "$filename" == "index" ]];
+  then
+    continue;
+  fi;
+
   rm index.js
   cp $file index.js
   zip $filename.zip index.js;
